# docs/openapi/backend.json
{
  "components": {
    "schemas": {
...
      "ConnectionBody": {
        "properties": {
          "connection": {
            "additionalProperties": true,
            "title": "Connection",
            "type": "object"
          }
        },
        "required": [
          "connection"
        ],
        "title": "ConnectionBody",
        "type": "object"
      },
      "ConnectionTestResponse": {
        "additionalProperties": true,
        "description": "``schemas`` is empty for connectors that cannot enumerate schemas.",
        "properties": {
          "schemas": {
            "items": {
              "type": "string"
            },
            "title": "Schemas",
            "type": "array"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "schemas"
        ],
        "title": "ConnectionTestResponse",
        "type": "object"
      },
...
      "DataResponse_PublicConnection_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/PublicConnection"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[PublicConnection]",
        "type": "object"
      },
...
      "DataResponse_SsoFederationState_": {
        "additionalProperties": true,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SsoFederationState"
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[SsoFederationState]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ListResponse_PublicConnection_": {
        "additionalProperties": true,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/PublicConnection"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data",
          "count"
        ],
        "title": "ListResponse[PublicConnection]",
        "type": "object"
      },
...
      "PublicConnection": {
        "additionalProperties": true,
        "description": "A UI-managed connection with its credentials stripped.\n\n``connection`` is absent from the delete response, which echoes only the\ndatabase name it tore down; the list and create reads carry the full\ncredential-free config.",
        "properties": {
          "connection": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connection"
          },
          "database_name": {
            "title": "Database Name",
            "type": "string"
          }
        },
        "required": [
          "database_name"
        ],
        "title": "PublicConnection",
        "type": "object"
      },
...
      "SsoFederationBody": {
        "properties": {
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          }
        },
        "required": [
          "enabled"
        ],
        "title": "SsoFederationBody",
        "type": "object"
      },
      "SsoFederationState": {
        "additionalProperties": true,
        "description": "The one flag ``PATCH /connections/{database_name}/sso-federation`` sets.\n\nThe route deliberately echoes just the flag rather than the connection, so\na caller never has to re-send credentials to toggle it.",
        "properties": {
          "database_name": {
            "title": "Database Name",
            "type": "string"
          },
          "sso_federation": {
            "title": "Sso Federation",
            "type": "boolean"
          }
        },
        "required": [
          "database_name",
          "sso_federation"
        ],
        "title": "SsoFederationState",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Auto Ontology API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/connections": {
      "get": {
        "description": "List every configured database connection, credential-free.\n\n``password`` and ``password_env`` are stripped from each connection object,\nso the payload is safe to render in the UI. ``database_name`` doubles as the\nconnection's identity in the other routes here.",
        "operationId": "list_connections_api_connections_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_PublicConnection_"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Connections",
        "tags": [
          "connections"
        ]
      },
      "post": {
        "description": "Register a database connection and return it credential-free.\n\nThe connection is validated on the way in; anything the driver rejects comes\nback as 422 with the failure message rather than a 500. Call\n``POST /connections/test`` first to check credentials and list the schemas\navailable for the picker.",
        "operationId": "create_connection_api_connections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_PublicConnection_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Connection",
        "tags": [
          "connections"
        ]
      }
    },
...
    "/api/connections/test": {
      "post": {
        "description": "Validate a connection and return its schemas (for the schema picker).",
        "operationId": "test_connection_api_connections_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionTestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Connection",
        "tags": [
          "connections"
        ]
      }
    },
    "/api/connections/{database_name}": {
      "delete": {
        "description": "Remove a connection, returning the deleted row.\n\n404 when no connection carries that ``database_name``. Data already ingested\nfrom the connection is not removed here \u2014 reset it through the ingestion\nservice's ``POST /ingest/delete``.",
        "operationId": "delete_connection_api_connections__database_name__delete",
        "parameters": [
          {
            "description": "Catalog database name, which doubles as the connection's identity.",
            "in": "path",
            "name": "database_name",
            "required": true,
            "schema": {
              "description": "Catalog database name, which doubles as the connection's identity.",
              "title": "Database Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_PublicConnection_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Connection",
        "tags": [
          "connections"
        ]
      }
    },
    "/api/connections/{database_name}/sso-federation": {
      "patch": {
        "description": "Toggle whether chat runs this connection's SQL as the signed-in user.",
        "operationId": "set_sso_federation_api_connections__database_name__sso_federation_patch",
        "parameters": [
          {
            "description": "Catalog database name, which doubles as the connection's identity.",
            "in": "path",
            "name": "database_name",
            "required": true,
            "schema": {
              "description": "Catalog database name, which doubles as the connection's identity.",
              "title": "Database Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SsoFederationBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_SsoFederationState_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Sso Federation",
        "tags": [
          "connections"
        ]
      }
    },
...
  }
}